[fit] 将handler修改为责任链形式
